refactor(controller): put controller code together
