--- plb/engine/primitive/__init__.py
@@ -1,1 +1,0 @@
-from .primitives import Primitives
